+++ packages/tailwindcss/src/compat/plugin-api.test.ts
@@ -1734,7 +1734,7 @@ describe('addVariant', () => {
 
     expect(optimizeCss(compiled).trim()).toMatchInlineSnapshot(`
       "@layer utilities {
-        @media (width <= 400px) {
+        @media (max-width: 400px) {
           @supports (font: bold) {
             .potato\\:flex:large-potato {
               display: flex;
@@ -2053,19 +2053,19 @@ describe('matchVariant', () => {
 
     expect(optimizeCss(compiled).trim()).toMatchInlineSnapshot(`
       "@layer utilities {
-        @media (width >= 500px) {
+        @media (min-width: 500px) {
           .testmin-\\[500px\\]\\:underline {
             text-decoration-line: underline;
           }
         }
 
-        @media (width >= 600px) {
+        @media (min-width: 600px) {
           .testmin-\\[600px\\]\\:flex {
             display: flex;
           }
         }
 
-        @media (width >= 700px) {
+        @media (min-width: 700px) {
           .testmin-\\[700px\\]\\:italic {
             font-style: italic;
           }
@@ -2108,19 +2108,19 @@ describe('matchVariant', () => {
 
     expect(optimizeCss(compiled).trim()).toMatchInlineSnapshot(`
       "@layer utilities {
-        @media (width >= 500px) {
+        @media (min-width: 500px) {
           .testmin-\\[500px\\]\\:italic {
             font-style: italic;
           }
         }
 
-        @media (width >= 600px) {
+        @media (min-width: 600px) {
           .testmin-example\\:italic {
             font-style: italic;
           }
         }
 
-        @media (width >= 700px) {
+        @media (min-width: 700px) {
           .testmin-\\[700px\\]\\:italic {
             font-style: italic;
           }
@@ -2168,30 +2168,30 @@ describe('matchVariant', () => {
 
     expect(optimizeCss(compiled).trim()).toMatchInlineSnapshot(`
       "@layer utilities {
-        @media (width >= 100px) {
-          @media (width <= 400px) {
+        @media (min-width: 100px) {
+          @media (max-width: 400px) {
             .testmin-\\[100px\\]\\:testmax-\\[400px\\]\\:order-1 {
               order: 1;
             }
           }
         }
 
-        @media (width >= 150px) {
-          @media (width <= 400px) {
+        @media (min-width: 150px) {
+          @media (max-width: 400px) {
             .testmin-\\[150px\\]\\:testmax-\\[400px\\]\\:order-2 {
               order: 2;
             }
           }
         }
 
-        @media (width >= 100px) {
-          @media (width <= 350px) {
+        @media (min-width: 100px) {
+          @media (max-width: 350px) {
             .testmin-\\[100px\\]\\:testmax-\\[350px\\]\\:order-3 {
               order: 3;
             }
           }
 
-          @media (width <= 300px) {
+          @media (max-width: 300px) {
             .testmin-\\[100px\\]\\:testmax-\\[300px\\]\\:order-4 {
               order: 4;
             }
@@ -2238,8 +2238,8 @@ describe('matchVariant', () => {
     // Expect :focus to come after :hover
     expect(optimizeCss(compiled).trim()).toMatchInlineSnapshot(`
       "@layer utilities {
-        @media (width >= 100px) {
-          @media (width <= 200px) {
+        @media (min-width: 100px) {
+          @media (max-width: 200px) {
             @media (hover: hover) {
               .testmin-\\[100px\\]\\:testmax-\\[200px\\]\\:hover\\:underline:hover {
                 text-decoration-line: underline;
@@ -2292,32 +2292,32 @@ describe('matchVariant', () => {
 
     expect(optimizeCss(compiled).trim()).toMatchInlineSnapshot(`
       "@layer utilities {
-        @media (width >= 100px) {
-          @media (width <= 400px) {
+        @media (min-width: 100px) {
+          @media (max-width: 400px) {
             .testmin-\\[100px\\]\\:testmax-\\[400px\\]\\:order-1 {
               order: 1;
             }
           }
         }
 
-        @media (width >= 200px) {
-          @media (width <= 400px) {
+        @media (min-width: 200px) {
+          @media (max-width: 400px) {
             .testmin-\\[200px\\]\\:testmax-\\[400px\\]\\:order-2 {
               order: 2;
             }
           }
         }
 
-        @media (width >= 100px) {
-          @media (width <= 300px) {
+        @media (min-width: 100px) {
+          @media (max-width: 300px) {
             .testmin-\\[100px\\]\\:testmax-\\[300px\\]\\:order-3 {
               order: 3;
             }
           }
         }
 
-        @media (width >= 200px) {
-          @media (width <= 300px) {
+        @media (min-width: 200px) {
+          @media (max-width: 300px) {
             .testmin-\\[200px\\]\\:testmax-\\[300px\\]\\:order-4 {
               order: 4;
             }
@@ -2364,28 +2364,28 @@ describe('matchVariant', () => {
 
     expect(optimizeCss(compiled).trim()).toMatchInlineSnapshot(`
       "@layer utilities {
-        @media (width <= 400px) {
-          @media (width >= 100px) {
+        @media (max-width: 400px) {
+          @media (min-width: 100px) {
             .testmax-\\[400px\\]\\:testmin-\\[100px\\]\\:underline {
               text-decoration-line: underline;
             }
           }
 
-          @media (width >= 200px) {
+          @media (min-width: 200px) {
             .testmax-\\[400px\\]\\:testmin-\\[200px\\]\\:underline {
               text-decoration-line: underline;
             }
           }
         }
 
-        @media (width <= 300px) {
-          @media (width >= 100px) {
+        @media (max-width: 300px) {
+          @media (min-width: 100px) {
             .testmax-\\[300px\\]\\:testmin-\\[100px\\]\\:underline {
               text-decoration-line: underline;
             }
           }
 
-          @media (width >= 200px) {
+          @media (min-width: 200px) {
             .testmax-\\[300px\\]\\:testmin-\\[200px\\]\\:underline {
               text-decoration-line: underline;
             }
@@ -2440,32 +2440,32 @@ describe('matchVariant', () => {
 
     expect(optimizeCss(compiled).trim()).toMatchInlineSnapshot(`
       "@layer utilities {
-        @media (width >= 100px) {
-          @media (width <= 400px) {
+        @media (min-width: 100px) {
+          @media (max-width: 400px) {
             .testmin-\\[100px\\]\\:testmax-\\[400px\\]\\:order-1 {
               order: 1;
             }
           }
         }
 
-        @media (width >= 200px) {
-          @media (width <= 400px) {
+        @media (min-width: 200px) {
+          @media (max-width: 400px) {
             .testmin-\\[200px\\]\\:testmax-\\[400px\\]\\:order-2 {
               order: 2;
             }
           }
         }
 
-        @media (width >= 100px) {
-          @media (width <= 300px) {
+        @media (min-width: 100px) {
+          @media (max-width: 300px) {
             .testmin-\\[100px\\]\\:testmax-\\[300px\\]\\:order-3 {
               order: 3;
             }
           }
         }
 
-        @media (width >= 200px) {
-          @media (width <= 300px) {
+        @media (min-width: 200px) {
+          @media (max-width: 300px) {
             .testmin-\\[200px\\]\\:testmax-\\[300px\\]\\:order-4 {
               order: 4;
             }
@@ -2629,13 +2629,13 @@ describe('matchVariant', () => {
       'my-container-[250px]/placement:underline',
     ])
     expect(optimizeCss(compiled).trim()).toMatchInlineSnapshot(`
-      "@container (width >= 250px) {
+      "@container (min-width: 250px) {
         .my-container-\\[250px\\]\\:underline {
           text-decoration-line: underline;
         }
       }
 
-      @container placement (width >= 250px) {
+      @container placement (min-width: 250px) {
         .my-container-\\[250px\\]\\/placement\\:underline {
           text-decoration-line: underline;
         }
@@ -2676,20 +2676,20 @@ describe('addUtilities()', () => {
 
     expect(optimizeCss(compiled.build(['text-trim', 'lg:text-trim'])).trim())
       .toMatchInlineSnapshot(`
-      "@layer utilities {
-        .text-trim {
-          text-box-trim: both;
-          text-box-edge: cap alphabetic;
-        }
-
-        @media (width >= 1024px) {
-          .lg\\:text-trim {
+        "@layer utilities {
+          .text-trim {
             text-box-trim: both;
             text-box-edge: cap alphabetic;
           }
-        }
-      }"
-    `)
+
+          @media (min-width: 1024px) {
+            .lg\\:text-trim {
+              text-box-trim: both;
+              text-box-edge: cap alphabetic;
+            }
+          }
+        }"
+      `)
   })
 
   test('return multiple rule objects from a custom utility', async () => {
@@ -2873,7 +2873,7 @@ describe('addUtilities()', () => {
           }
         }
 
-        @media (width >= 1024px) {
+        @media (min-width: 1024px) {
           .lg\\:foo {
             display: flex;
           }
@@ -2954,7 +2954,7 @@ describe('addUtilities()', () => {
           background-color: #fff;
         }
 
-        @media (width >= 1024px) {
+        @media (min-width: 1024px) {
           .lg\\:form-textarea {
             appearance: none;
             background-color: #fff;
@@ -3280,7 +3280,7 @@ describe('matchUtilities()', () => {
         border-block-width: var(--foo);
       }
 
-      @media (width >= 1024px) {
+      @media (min-width: 1024px) {
         .lg\\:border-block-2 {
           border-block-width: 2px;
         }
@@ -3579,7 +3579,7 @@ describe('matchUtilities()', () => {
         }
 
         .scrollbar-\\[\\#08c\\]\\/50 {
-          scrollbar-color: oklab(59.9824% -.06725 -.12414 / .5);
+          scrollbar-color: oklab(59.9824% -.067 -.124 / .5);
         }
 
         .scrollbar-\\[2px\\] {
@@ -3975,7 +3975,7 @@ describe('matchUtilities()', () => {
           display: flex;
         }
 
-        @media (width >= 1024px) {
+        @media (min-width: 1024px) {
           .lg\\:foo-\\[12px\\] {
             --foo: 12px;
             display: flex;