+++ packages/tailwindcss/src/utilities.ts
@@ -2,6 +2,7 @@ import {
   atRoot,
   atRule,
   decl,
+  rule,
   styleRule,
   walk,
   type AstNode,
@@ -185,9 +186,7 @@ export function withAlpha(value: string, alpha: string): string {
 /**
  * Apply opacity to a color using `color-mix`.
  */
-export function replaceAlpha(value: string, alpha: string | null): string {
-  if (alpha === null) return value
-
+export function replaceAlpha(value: string, alpha: string): string {
   // Convert numeric values (like `0.5`) to percentages (like `50%`) so they
   // work properly with `color-mix`. Assume anything that isn't a number is
   // safe to pass through as-is, like `var(--my-opacity)`.
@@ -2039,6 +2038,9 @@ export function createUtilities(theme: Theme) {
   }
   staticUtility('self-stretch', [['align-self', 'stretch']])
   staticUtility('self-baseline', [['align-self', 'baseline']])
+  if (enableBaselineLast) {
+    staticUtility('self-baseline-last', [['align-self', 'last baseline']])
+  }
 
   staticUtility('justify-self-auto', [['justify-self', 'auto']])
   staticUtility('justify-self-start', [['justify-self', 'flex-start']])
@@ -2533,7 +2535,10 @@ export function createUtilities(theme: Theme) {
         let interpolationMethod = resolveInterpolationModifier(candidate.modifier)
 
         return [
-          decl('--tw-gradient-position', `${value} ${interpolationMethod}`),
+          decl('--tw-gradient-position', `${value}`),
+          rule('@supports (background-image: linear-gradient(in lab, red, red))', [
+            decl('--tw-gradient-position', `${value} ${interpolationMethod}`),
+          ]),
           decl('background-image', `linear-gradient(var(--tw-gradient-stops))`),
         ]
       }
@@ -3639,15 +3644,15 @@ export function createUtilities(theme: Theme) {
   staticUtility('object-none', [['object-fit', 'none']])
   staticUtility('object-scale-down', [['object-fit', 'scale-down']])
 
-  staticUtility('object-bottom', [['object-position', 'bottom']])
-  staticUtility('object-center', [['object-position', 'center']])
-  staticUtility('object-left', [['object-position', 'left']])
-  staticUtility('object-left-bottom', [['object-position', 'left bottom']])
-  staticUtility('object-left-top', [['object-position', 'left top']])
-  staticUtility('object-right', [['object-position', 'right']])
-  staticUtility('object-right-bottom', [['object-position', 'right bottom']])
-  staticUtility('object-right-top', [['object-position', 'right top']])
   staticUtility('object-top', [['object-position', 'top']])
+  staticUtility('object-top-left', [['object-position', 'left top']])
+  staticUtility('object-top-right', [['object-position', 'right top']])
+  staticUtility('object-bottom', [['object-position', 'bottom']])
+  staticUtility('object-bottom-left', [['object-position', 'left bottom']])
+  staticUtility('object-bottom-right', [['object-position', 'right bottom']])
+  staticUtility('object-left', [['object-position', 'left']])
+  staticUtility('object-right', [['object-position', 'right']])
+  staticUtility('object-center', [['object-position', 'center']])
   functionalUtility('object', {
     themeKeys: ['--object-position'],
     handle: (value) => [decl('object-position', value)],
@@ -5140,9 +5145,11 @@ export function createUtilities(theme: Theme) {
         return [
           textShadowProperties(),
           decl('--tw-text-shadow-alpha', alpha),
-          decl(
+          ...alphaReplacedShadowProperties(
             'text-shadow',
-            replaceShadowColors(value, alpha, (color) => `var(--tw-text-shadow-color, ${color})`),
+            value,
+            alpha,
+            (color) => `var(--tw-text-shadow-color, ${color})`,
           ),
         ]
       }
@@ -5164,13 +5171,11 @@ export function createUtilities(theme: Theme) {
             return [
               textShadowProperties(),
               decl('--tw-text-shadow-alpha', alpha),
-              decl(
+              ...alphaReplacedShadowProperties(
                 'text-shadow',
-                replaceShadowColors(
-                  value,
-                  alpha,
-                  (color) => `var(--tw-text-shadow-color, ${color})`,
-                ),
+                value,
+                alpha,
+                (color) => `var(--tw-text-shadow-color, ${color})`,
               ),
             ]
           }
@@ -5190,9 +5195,11 @@ export function createUtilities(theme: Theme) {
           return [
             textShadowProperties(),
             decl('--tw-text-shadow-alpha', alpha),
-            decl(
+            ...alphaReplacedShadowProperties(
               'text-shadow',
-              replaceShadowColors(value, alpha, (color) => `var(--tw-text-shadow-color, ${color})`),
+              value,
+              alpha,
+              (color) => `var(--tw-text-shadow-color, ${color})`,
             ),
           ]
         }
@@ -5280,9 +5287,11 @@ export function createUtilities(theme: Theme) {
         return [
           boxShadowProperties(),
           decl('--tw-shadow-alpha', alpha),
-          decl(
+          ...alphaReplacedShadowProperties(
             '--tw-shadow',
-            replaceShadowColors(value, alpha, (color) => `var(--tw-shadow-color, ${color})`),
+            value,
+            alpha,
+            (color) => `var(--tw-shadow-color, ${color})`,
           ),
           decl('box-shadow', cssBoxShadowValue),
         ]
@@ -5306,9 +5315,11 @@ export function createUtilities(theme: Theme) {
             return [
               boxShadowProperties(),
               decl('--tw-shadow-alpha', alpha),
-              decl(
+              ...alphaReplacedShadowProperties(
                 '--tw-shadow',
-                replaceShadowColors(value, alpha, (color) => `var(--tw-shadow-color, ${color})`),
+                value,
+                alpha,
+                (color) => `var(--tw-shadow-color, ${color})`,
               ),
               decl('box-shadow', cssBoxShadowValue),
             ]
@@ -5333,9 +5344,11 @@ export function createUtilities(theme: Theme) {
           return [
             boxShadowProperties(),
             decl('--tw-shadow-alpha', alpha),
-            decl(
+            ...alphaReplacedShadowProperties(
               '--tw-shadow',
-              replaceShadowColors(value, alpha, (color) => `var(--tw-shadow-color, ${color})`),
+              value,
+              alpha,
+              (color) => `var(--tw-shadow-color, ${color})`,
             ),
             decl('box-shadow', cssBoxShadowValue),
           ]
@@ -5395,9 +5408,11 @@ export function createUtilities(theme: Theme) {
         return [
           boxShadowProperties(),
           decl('--tw-inset-shadow-alpha', alpha),
-          decl(
+          ...alphaReplacedShadowProperties(
             '--tw-inset-shadow',
-            replaceShadowColors(value, alpha, (color) => `var(--tw-inset-shadow-color, ${color})`),
+            value,
+            alpha,
+            (color) => `var(--tw-inset-shadow-color, ${color})`,
           ),
           decl('box-shadow', cssBoxShadowValue),
         ]
@@ -5421,13 +5436,12 @@ export function createUtilities(theme: Theme) {
             return [
               boxShadowProperties(),
               decl('--tw-inset-shadow-alpha', alpha),
-              decl(
+              ...alphaReplacedShadowProperties(
                 '--tw-inset-shadow',
-                `inset ${replaceShadowColors(
-                  value,
-                  alpha,
-                  (color) => `var(--tw-inset-shadow-color, ${color})`,
-                )}`,
+                value,
+                alpha,
+                (color) => `var(--tw-inset-shadow-color, ${color})`,
+                'inset ',
               ),
               decl('box-shadow', cssBoxShadowValue),
             ]
@@ -5453,13 +5467,11 @@ export function createUtilities(theme: Theme) {
           return [
             boxShadowProperties(),
             decl('--tw-inset-shadow-alpha', alpha),
-            decl(
+            ...alphaReplacedShadowProperties(
               '--tw-inset-shadow',
-              replaceShadowColors(
-                value,
-                alpha,
-                (color) => `var(--tw-inset-shadow-color, ${color})`,
-              ),
+              value,
+              alpha,
+              (color) => `var(--tw-inset-shadow-color, ${color})`,
             ),
             decl('box-shadow', cssBoxShadowValue),
           ]
@@ -6208,5 +6220,44 @@ function resolveValueFunction(
         return { nodes: ValueParser.parse(value.value) }
       }
     }
+  }
+}
+
+function alphaReplacedShadowProperties(
+  property: string,
+  value: string,
+  alpha: string | null | undefined,
+  varInjector: (color: string) => string,
+  prefix: string = '',
+): AstNode[] {
+  let requiresFallback = false
+  let replacedValue = replaceShadowColors(value, (color) => {
+    if (alpha == null) {
+      return varInjector(color)
+    }
+
+    // When the input is currentcolor, we use our existing `color-mix(…)` approach to increase
+    // browser support. Note that the fallback of this is handled more generically in
+    // post-processing.
+    if (color.startsWith('current')) {
+      return varInjector(withAlpha(color, alpha))
+    }
+
+    // If any dynamic values are needed for the relative color syntax, we need to insert a
+    // replacement as lightningcss won't be able to resolve them statically.
+    if (color.startsWith('var(') || alpha.startsWith('var(')) {
+      requiresFallback = true
+    }
+
+    return varInjector(replaceAlpha(color, alpha))
+  })
+
+  if (requiresFallback) {
+    return [
+      decl(property, prefix + replaceShadowColors(value, varInjector)),
+      rule('@supports (color: lab(from red l a b))', [decl(property, prefix + replacedValue)]),
+    ]
+  } else {
+    return [decl(property, prefix + replacedValue)]
   }
 }